Move sum methods to new interface (SOLID)
